v3 - move extension folders to ext
