+++ ext/prompt/status_string.go
